+++ src/communication/azext_communication/tests/latest/recording_processors.py
@@ -26,6 +26,8 @@ class URIIdentityReplacer(RecordingProcessor):
         request.uri = re.sub('/identities/([^/?]+)', '/identities/sanitized', request.uri)
         request.uri = re.sub('/chat/threads/([^/?]+)', '/chat/threads/sanitized', request.uri)
         request.uri = re.sub('/chat/threads/([^/?]+)/messages/([^/?]+)', '/chat/threads/sanitized/messages/sanitized', request.uri)
+        request.uri = re.sub('/rooms/([0-9]+)/', '/rooms/sanitized/', request.uri)
+        request.uri = re.sub('/rooms/([0-9]+)\?', '/rooms/sanitized?', request.uri)
         request.uri = re.sub(resource, 'sanitized', request.uri)
         return request
     
@@ -35,6 +37,7 @@ class URIIdentityReplacer(RecordingProcessor):
             response['url'] = re.sub('/identities/([^/?]+)', '/identities/sanitized', response['url'])
             response['url'] = re.sub('/chat/threads/([^/?]+)', '/chat/threads/sanitized', response['url'])
             response['url'] = re.sub('/chat/threads/([^/?]+)/messages/([^/?]+)', '/chat/threads/sanitized/messages/sanitized', response['url'])
+            response['url'] = re.sub('/rooms/([0-9]+)/', '/rooms/sanitized/', response['url'])
         return response
 
 
@@ -134,13 +137,21 @@ class BodyReplacerProcessor(RecordingProcessor):
         return response
     
     def _replace_keys(self, body):
-        def _replace_recursively(dictionary):
-            for key in dictionary:
-                value = dictionary[key]
-                if key in self._keys:
-                    dictionary[key] = self._replacement
-                elif isinstance(value, dict):
-                    _replace_recursively(value)
+        import collections.abc
+        def _replace_recursively(data):
+            if (isinstance(data, str)):
+                return
+            
+            if isinstance(data, dict):
+                for key in data:
+                    value = data[key]
+                    if key in self._keys:
+                        data[key] = self._replacement
+                    else:
+                        _replace_recursively(value)
+            elif isinstance(data, collections.abc.Sequence):
+                for item in data:
+                    _replace_recursively(item)
 
         import json
         try: