+++ src/communication/HISTORY.rst
@@ -3,10 +3,27 @@
 Release History
 ===============
 
+1.4.0
+++++++
+ * Add communication rooms command group in preview mode
+ * Add confirmation for delete/remote/revoke commands under identity, chat, rooms
+ * Update minCliCoreVersion to 2.40.0
+
+
+1.3.0
+++++++
+ * Add AzureCli to the user-agent header
+
+
+1.2.2
+++++++
+ * Fix a bug in chat 'message delete command'
+
+
 1.2.1
 ++++++
  * Update command helps
- * Fix bugs preview and deprecated flags
+ * Fix bugs in preview and deprecated flags
 
 
 1.2.0